libtracker-miner: Add ownership tracking to Tracker indexing config
The TrackerIndexingTree class keeps track of each directory or file that
has been explicitly selected for indexing and monitoring.

As a new feature, we want to allow any application to add to the list of
directories being indexed and monitored, and to later be able to remove
things from the list again. Previously, the tracker-miner-fs process was
totally in control of what went in the IndexingTree.

In order to ensure that applications can't interfere with the
existing configuration, we need to track *who* added something to the
indexing tree. This is done with a string identifier.

The tracker_miner_fs_directory_add() functions now take an 'owner'
parameter so they can keep track of who is interested in each configured
indexing root. The tracker_miner_fs_ignore() function lets a given owner
that previously expressed interest in an indexing root now say that they
are not interested. If that indexing root has no 'owners' left, it is
deleted from the TrackerIndexingTree, which cancels any running indexing
tasks and removes any monitors.

We need to make sure that nobody can claim the internal
'tracker-miner-fs' ID through the external API -- that would make it
possible to override the internal configuration.