libtracker-miner: Add ownership tracking to TrackerIndexingTree
The TrackerIndexingTree class keeps track of each directory or file that
has been explicitly selected for indexing and monitoring.

As a new feature, we want to allow any application to add to the list of
directories being indexed and monitored, and to later be able to remove
things from the list again. Previously, the tracker-miner-fs process was
totally in control of what went in the IndexingTree.

In order to ensure that applications can't interfere with the
existing configuration, we need to track *who* added something to the
indexing tree. This is done with a string identifier.

We need to make sure that nobody can claim the internal
'tracker-miner-fs' ID through the external API -- that would make it
possible to override the internal configuration.

FIXME: STILL NEED TO WRITE SOME TESTS FOR THIS !!