libtracker-miner: Add ownership tracking, and tracker_miner_fs_ignore()
This lets us add external (D-Bus) functions to control the internal
TrackerIndexingTree, and thus control from outside the miner what is
being indexed and what is not. Previously the only way to control that
was via GSettings.

The tracker_miner_fs_directory_add() functions now take an 'owner'
parameter so they can keep track of who is interested in each configured
indexing root. The tracker_miner_fs_ignore() function lets a given owner
that previously expressed interest in an indexing root now say that they
are not interested. If that indexing root has no 'owners' left, it is
deleted from the TrackerIndexingTree, which cancels any running indexing
tasks and removes any monitors.