tracker-miner-fs: Keep cache of IndexFile requesters on directories
The senders for all DBus requests to the IndexFile method on directories
will be now watched, the dbus presence of the senders will control the
lifetime of the directory on the indexed directories tree.

There may be multiple requests on a same directory, in such case the
directory will be indexed/monitored for as long as there is alive requesters
on it. Requests on already indexed directories (or children of recursively
indexed ones) will be silently ignored. Unmounts will also silently drop
the IndexFile listeners, applications should issue new requests on volume
mounts if desired.

The patch is loosely based on initial work from Felipe Borges.

https://bugzilla.gnome.org/show_bug.cgi?id=680834