libtracker-extract: fix dummy_module
The ModuleInfo typedef contains 6 items, of which the last is 'optional' (predefined).

TRUE as parameter for TrackerExtractShutdownFunc though is not a good idea: it's used as an
integer and passed as a pointer to a function call.

As there is no shudown function for the dummy module, just pass an additional NULL in there.

https://bugzilla.gnome.org/show_bug.cgi?id=754448