Add systemd user services corresponding to D-Bus session services
The one exception is the email miner, which does not actually have
anything to start, so its D-Bus service file won't work anyway.

Restart the services on failure, but do not after SIGKILL,
because that's what `tracker daemon --kill` (formerly
`tracker-control --kill`) does. systemd already treats SIGTERM
as a graceful exit which is not considered to be a failure.

Loosely based on a patch by Alban Crequy, which only applied
this to tracker-store and tracker-miner-fs, and did not have
the special handling for SIGKILL.

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugzilla.gnome.org/show_bug.cgi?id=755739
Reviewed-by: Carlos Garnacho