libtracker-data: Resolve variables looking up the binding if we're in same context
This is necessary for BIND support, as the variables looked up to add the
BIND alias likely belong to the same scope that we're currently building.
This results in SQL like (very simplified, and shortened):

  SELECT "nie:url" AS "1_u", "1_u" AS "2_u" ...

The second "1_u" can't be looked up yet in this scope (only on the WHERE
clause, or a wrapping SELECT), so ends up as the string itself. With this
change, this kind of queries will result in (shortened again):

  SELECT "nie:url" AS "1_u", "nie:url" AS "2_u" ...

Which is legal, and results in the expected values assigned to the BIND
alias.