libtracker-data: Add support for BIND
The BIND form allows a value to be assigned to a variable from a graph
pattern:

  SELECT ?u ?hour
  { ?u a rdfs:Resource .
       nfo:fileLastModified ?modified
    BIND (HOUR (?modified) AS ?hour)
  }

It will be specially useful on places where we use the result of a
calculation on more than one places, eg. on SELECT and FILTER(). It will
also be useful combined with DELETE...INSERT...WHERE to perform complex
mass-updates in a single query.

http://www.w3.org/TR/sparql11-query/#bind