libtracker-miner: Optimize move operations
Instead of creating one delete+insert per file (indirectly) contained
in a folder, do it all at once in a DELETE...INSERT...WHERE. We only
now need to query and block if we're running a thumbnailer service,
which should be most uncommon.

This makes the whole update to happen in a much tighter loop within
tracker-store, eg. helps reduce the time spent in processing the
renaming of a linux kernel tree (51964 elements) from 23s to 12s.
(as measured locally by tracker-miner-fs on a previously idle
system).