miners: Take over the deletion of previous properties from TrackerMinerFS
Performing the deletion inside libtracker-miner takes some assumptions we
shouldn't be doing in a generic library, such as the graph we're deleting
from. It makes more sense to make miners responsible of deleting their
older properties prior to a file update.

Applications/user guides miners could be more specific as to which
properties they're deleting, we'll rely on those not stepping over other
miners' toes at the moment.

https://bugzilla.gnome.org/show_bug.cgi?id=758341