libtracker-miner: Do not insert partial/empty sparql on error
The check for these errors was done specifically so we could still
insert (even if incomplete) data on tracker-extract failures, when
we used to communicate with it directly from tracker-miner-fs.

Nowadays, tracker-extract is a TrackerDecorator, and tracker-miner-fs
should most likely receive only errors here on ENOENT and other errors
that affect the file and its info as a whole. In these situations
we end up with a task with a completely empty sparql string, which
doesn't help much here.