libtracker-miner: Do not break ordering with bulk operations
The way bulk operations break the ordering of the sparql updates
may bring certain glitches. eg. if a delete (bulk) operation
happens on a folder while the sparql buffer already contains
updates corresponding to files within that folder, the way we
order operations results in the delete happening first, and
the inserts/updates happening afterwards. This may leave the
database in inconsistent states.

So make bulk operations a 1 file thing, this could maybe be
smarter and try to compress consecutive similar operations,
but there's not much added value in it. Plus, the bulk operations
we issue in the miner would never match that criteria.

Because we no longer break the ordering with bulk operations,
things like the error map are no longer necessary, so we can
remove some icky code here.

BulkOperationMerge and its functions could be further simplified,
that's left for a later cleanup.