libtracker-direct: Handle cancellation
Always go through the async path, and manage the cancellable in
both the IO scheduler and the jobs. If we try to handle cancellation
in the direct path, the cursor will be unref'ed (and try to grab
the DB lock again) while the lock is held. There's no way cancellation
can work while holding the lock, so just removed this path.