libtracker-miner: Avoid querying file type in crawling queries
We only really need the file type from the database when checking
for contents that are no longer in the filesystem. This spares us
the OPTIONAL in the (by far) most commonly performed query here,
the one checking the contents that do exist in the filesystem.

Also, modify the "deleted contents check" query so it performs
better with a larger number of elements.