tracker-extract: Fix gstreamer module cuesheet handling
When a single nfo:FileDataObject represents multiple nmm:MusicPieces
(eg. flac files with a cue sheet). We attempted to drop the nmm:MusicPiece
aspect from the nfo:FileDataObject, but we dropped the entire resource
instead.

What was intended here is only dropping the nmm:MusicPiece, so the file
resource remains a nfo:Audio and nfo:FileDataObject, and the per-track
nmm:MusicPiece backreference to it through nie:isStoredAs.