IPTC extractors : use nco:contributor for contact metadata
5a27279a7e07a9f1f0c58f663f4c290c397d4aff adds support for contact
IPTC metadata and represents it with a nco:representative predicate.

This contradicts the ontology that stipulates that the subject of
a nco:representative predicate is a nco:Contact. Consequently, updates
from extraction of files containing this IPTC field are rejected by
the store.

This fix uses the nco:contributor predicate instead. Official
documentation at [1] says the field "Identifies the person or
organisation which can provide further background information on
the objectdata." Nco:contributor is therefore too narrow as not every
such person need be a contributor. The adequate predicate doesn't
exist but if it did it would definitely at least be a nao:annotation
or more narrowly a nao:isRelated. But I deem these too broad because
their range isn't a person or an organisation.

https://bugzilla.gnome.org/show_bug.cgi?id=762675