libtracker-data: Handle inserts where the subproperty cardinality is larger than the parents'
This is the case of properties like nco:creator, which has unrestricted
cardinality, but is subproperty of dc:creator, which has its cardinality
limited to 1.

In cases like this, inserts of multiple values shouldn't fail because of
the restricted cardinality in superproperties which this one supersedes.
So make it sure we don't propagate inserts to superproperties if its
cardinality is already satisfied.

Updates already work fine in this regard, with the difference that it's
the last value which will prevail in the single-valued superproperty, this
seems consistent with the nature of updates.