tracker-extract: Propagate urn to the TrackerExtractInfo
Now that extraction is performed after the file got initially
inserted and has an URN, there's no reason we can't pass the
URN to the TrackerExtractInfo. This would make URNs available
to extract modules, which have some places where the SPARQL
produced would be clearer.

This actually opens the path for quite a cleanup, we might
pass a single TrackerSparqlBuilder to extract modules instead
of doing the postprocessing we currently perform.