tracker-extract: Use safer method to insert tags in JPEG module
The current way of inserting the nao:hasTag relationship on the
extracted file involves one join operation per tag being inserted.
This has performance implications, plus we can feasibly hit the
sqlite limit of 64 tables in joins.

Instead insert the tags in separate inserts, that will be as fast
as it gets, plus there's no limit in the number of tags.