libtracker-miner: Ensure stored mtime data is freed after use
This data is not needed after comparing store and filesystem mtimes,
and keeping it stored could mislead later checks happening on that
same folder (say, directories being recursively rechecked because of
configuration changes).

It is safer to just dispose the data after use, the data might need
retrieving again at the time such recheck is needed.