libtracker-miner: Add indexing roots to the tree only when those are crawled
We must not insert information for indexing roots prior to their processing.
This could be the case of indexing roots that are descendants of another
recursive root (eg. ~/Music in ~/).

This was inocuous when crawling the indexed folders from scratch, although
could make indexing roots being spuriously deleted/reindexed when folders
are reindexed (eg. because of TrackerIndexingTree::directory-updated signals)