libtracker-miner: Do not free the current root data automatically on interruption
Most interruption cases are global to a RootData (eg. unmounts). It however
may turn out with unexpected consequences when requesting crawling of a
subfolder that is being deleted. In that case we'd expect the RootData to
be preserved and crawling to resume on the remaining folders.

So, only delete the current root data across regular operations as long
as we're not being interrupted, and ensure we do so manually in all the
cases where tracker_crawler_stop() is manually called.