libtracker-miner: Give up on detailed decorator error logging
It was the hope that this error logging would lead to useful
but reports that helped qualitatively improve the extractor
modules soon enough.

But people suck, and instead we got lots of bugs about how
much we spam syslog (as if we used the "warning" logging level
just because) or how we violated their privacy, without further
thinking that there might be reasons for those warnings in the
first place (i.e. bugs that need fixing). Those bug reports
are usually plagued with half-cut,half-edited snippets from
syslog to prove their point, while at the same time invalidating
their usefulness.

So, change of approach. Now we just include the URN/URI in the
error warning, and point people to using TRACKER_VERBOSITY=3 in
order to get more information. The debug information will contain
the full unedited sparql, and the relevant info will be surrounded
by dotted lines so it is made clear the information that is useful.

This will require an extra step for the people trying to be helpful,
which hopefully those are willing to take. It will also hopefully
calm the bunch of complains, and will give me no hesitation to close
any further one we get as invalid.