tracker-miner-fs: Add MS Office "owner files" to ignored-files
"Owner files" (described in https://support.microsoft.com/en-us/kb/211632)
works as a lock file when opening the document. Those are supposedly
removed when the document is closed, but in practice there's situations
where those are left behind, and are meaningless to tracker.

Those files may be generated for any of the document extensions, so handle
the whole bunch.

https://bugzilla.gnome.org/show_bug.cgi?id=764739