libtracker-common: Actually ignore XDG dirs set to $HOME
If an XDG dir is set to $HOME, we do ignore it (So it's not inadvertently
indexed recursively). However we just made the expansion to the real path
fail, the alias (e.g. &VIDEOS) was still dealt with as a path, and
attempted to be indexed.

So now we always return TRUE if the alias has a mapping, but only return
the string if the path is to our liking. This makes the alias truly
ignored.