docs: Do not delete xml directory when redoing docs
We're competing with gtk-doc.make here, which may turn out in
failures when doing mkdir on an already created directory, or
failures in writing xml/gtkdocentities.ent because we deleted
the directory under its feet.

So just do the same, use MKDIR_P to ensure it won't fail if
the directory is already there, and don't bother deleting it
as it's handled in gtk-doc.make. The ontology docs helper app
will rewrite the xml files anyway.