libtracker-data: Refine BIND() variable mapping
The implementation of BIND() introduced an special case in the way we
construct the result columns in selects (Handled in commit 64b54a7784).
This special case however turns out wrong when the resolved variable
is used both in and outside the BIND() statement.

So, admit the nature of this hack and special case it altogether, the
check was too iffy when there is really one situation where this makes
sense.

https://bugzilla.gnome.org/show_bug.cgi?id=766505