tracker-extract: Consider DjVu multipage docs as Ebooks
ddb79e3 and cf7f5df tried to categorise DjVu files as paginated
documents, which would have slotted them next to PDF files in
gnome-documents.

But we'd like them to be near EPubs in gnome-books. As the EBook RDF
type is not too strict in its definition, tagging the DjVu multi-page
files as Ebooks is the easy way to fix it.

https://bugzilla.gnome.org/show_bug.cgi?id=766781