fts: Do not apply stop-words when tokenizing query search terms
FTS5 notifies of the purpose in tokenization in their xTokenize vfunc,
check for the FTS5_TOKENIZE_QUERY flag indicating that this is the
tokenization of the search terms in a query and do not apply the stop
words list in this case.

One example where this is potentially harmful are "search as you type"
UIs. eg. typing the word "ontology" would have you type the word "onto",
which is an ignored word. Only after typing the next character you'd get
matches, which seems irregular behavior.