tracker: Fix possible warnings with the tracker search -[msivt] subcommands
We used to poke a third column when the "no search term given" queries just
return 2. This triggers warnings in the bus backend, but doesn't with the
direct backend, which just returns NULL.

Some people are seeing these warnings, which sounds bad enough because it
means tracker is forcing the "bus" backend for some reason. But of all ways
to notice this situation, this should not be one.