libtracker-miner: Move to next priority queue segment after deleting tail
If the tail element of a segment is being removed, we didn't move on to the
next one, so accounting of subsequent segments might get broken if further
elements are removed that are the head or the tail of a segment.

Also, just do the "jump to next segment" bit generic, so the affected paths
just turn on a fetch_segment boolean flag.

https://bugzilla.gnome.org/show_bug.cgi?id=769035