libtracker-data: Make TrackerDBStatement a GInitiallyUnowned
We actually deal with it as if it were (cached stmts have one ref
owned by the cache, and one by the caller. non-cached has just one
reference).

So make TrackerDBStatement actually an unowned type, and
g_object_ref_sink() in both places, if the stmt is cached the
ref will be sunk by the cache, and an extra ref added for the
caller, and if non-cached the only sunk reference will be the
callers'.