libtracker-miner: Remove all traces of libmediaart
We just happen to use it to remove related media art when files
are deleted. This should instead be deferred to a low urgency
maintenance task (and probably outside tracker, the albumart
dir is not just ours).

Anyway, tracker never got this entirely right, there can be
other copies of the files which would get their albumart
taken away of a sudden. Or we are simply not guaranteed to
receive file monitor events (eg. disabling the setting, or
due to monitor exhaustion).