libtracker-miner: Start TrackerDecorator queries on TrackerMiner::started
This should not be done in the initable vfunc, as the miner should remain
idle until tracker_miner_start() is called on it.

This fixes a possible race condition in tracker-extract, where this query
is already in flight at the time TrackerPersistence handles blacklisting
of files that caused earlier crashes of tracker-extract. So the query might
contain the file that's being blacklisted right away. tracker_miner_start()
is called after TrackerPersistence did the blacklisting job, so the first
query will get the up-to-date info.