libtracker-data: Handle overflows on libicu-based normalization
We allocate by default double the string size, which is a pretty
generous allotment most usually. If that is not enough for storing
the normalized string, just reallocate the buffer with the given
expected size and try again.

Also, refactor string normalization to a separate function, so
it's shared by both tracker:normalize and tracker:unaccent.

https://bugzilla.gnome.org/show_bug.cgi?id=769982