tracker-extract: Don't mix artist and albumartist
The 'album artist' is not the same thing as the 'performer' or the
'artist' metadata. In some extractors these tags however got mixed to
provide fallback values for certain tags. In the long run this is not a
viable approach: the tracker information should be a true representation
of what is in the tags.

https://bugzilla.gnome.org/show_bug.cgi?id=773697