libtracker-data: Use internal mutex in TrackerDBInterfaceSqlite
There is no reason to use a global mutex for multiple TrackerDBInterfaces,
the threading restriction in sqlite is "no multiple threads using the same
sqlite3* at the same time" (being cursor included "using" there). This
means we must restrict simultaneous threaded access to a single
TrackerDBInterface, but multiple threads are free to use a
TrackerDBInterface each without further interference.

So, use a per-TrackerDBInterface mutex instead of a global one, if multiple
cursors are created on the same interface, they'll backreference to it, so
will be blocking on the same mutex.