libtracker-data: Remove global interface
We no longer need to restrict to a single TrackerDBInterface for the
direct connection use. This also allows to maximize the throughput
in clients performing multiple simultaneous async queries, as we're
going from using a single TrackerDBInterface to satisfy all requests
(with the locking overhead this involves) to using multiple database
connections that can run different concurrent queries each.

This (and the last commits) translates to faster concurrent access,
which evolves close to linearly under much higher loads, instead of
exponentially. It is expected though that O(n^m) happens again if
the number of operations vastly exceed the number of interfaces
serving requests, but that's much farther.

As an example, the numbers when running multiple concurrent queries
for "SELECT ?u WHERE { ?u a rdfs:Resource }" (returning 12897
elements each in this computer) progress like this on an otherwise
idle session:

Simul. Queries	With		Without
5		    0,399s	    0,437s
10		    0,700s	    0,842s
50		    3,621s	    4,285s
100		    7,315s	    8,956s
500		   36,012s	1m 24,952s
1000		1m 16,355s      8m 33,880s

So as expected the numbers really shine when the thread contention
is most noticeable, and quite close to linearly (eg. in the "with"
column, every row is roughly 10x higher than the one corresponding
to simultaneous queries/10)