libtracker-direct: Replace TrackerDBManager lock usage with private one
We need to lock here not because of the TrackerDBInterface, but because
of other misc things needed at the time of constructing the SQL query
(like TrackerOntologies, which constructs data on demand from the gvdb,
this can't work across multiple threads).

But we don't need to use the TrackerDBManager lock, it can be just
a private one, and we remove another usage of this global lock.