libtracker-data: Add explicit calls to grab/release an stmt
An stmt can just be grabbed once at a time, and will ref both
the stmt and its DB interface for as long as the stmt is grabbed.
After releasing both refs will be dropped (and stmt_is_used set
back to FALSE), so it will be ensured that a database connection
lives as long as there are active cursors grabbing an statement.