libtracker-data: Handle BIND while in group graph pattern
Currently SQL construction fails while using the BIND form inside
a group graph pattern, eg. after OPTIONALs. When in that stage,
the OPTIONALs will be already resolved to SQL in the "FROM ..."
clause as nested SELECTs.

As such the generated SQL is wrong, make it handle this case
by wrapping the nested SELECTs under "SELECT $bind_value, * FROM (...)"
so the value is made available to the wrapping SELECT. The older
code still applies when building simple "{ ... . BIND (...) }" clauses.

https://bugzilla.gnome.org/show_bug.cgi?id=774251