libtracker-data: Return NULL SQL expression if the binding is not yet setup
Recent vala changed property setters so they compare with the previous
value before emitting anything, this triggers the property getter called
at a time when we just don't have a DataTable to construct the SQL
expression (and it's being set anyway).

So ensure the binding is in the right state before trying to construct a
SQL expression for it.