libtracker-data: Wrap BIND argument with () in SQL
Constructing the query as "SELECT $bind * FROM (...)" may trigger
sqlite parser overflows if the BIND form turns out to be too complex,
Doing the query as "SELECT ($bind) * FROM (...)" seems to be more
friendly to the parser.

This error was seen in gnome-music search, where the BIND() argument
is something like:

BIND((IF(STRSTARTS(?title_lower, "the "), SUBSTR(?title_lower, 5),
         IF(STRSTARTS(?title_lower, "a "), SUBSTR(?title_lower, 3),
            IF(STRSTARTS(?title_lower, "an "), SUBSTR(?title_lower, 4),
               fn:replace(fn:lower-case(?title_lower),
                          "^[ !\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~]+|[ !\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~]+$", "")
              )
           )
        )
     ) AS ?title_collation)

Which, despite the fn:replace regex argument being passed as a host
parameter, seemed too much to sqlite.