libtracker-sparql: Remove comment
It doesn't really apply. The DELETE DATA { pattern } query requires
the exact given triples, and won't work on variables. As it's not
nice or possible to query the current values just to perform a better
looking DELETE DATA {} query, just stick to the DELETE {} WHERE {}
one, and don't insinuate that we have a better option.