configure: Disable libmediaart dependency
Hopefully just temporarily. At the very least the processing
should be taken out of the extractor threads, as it requires
file operations that are not allowed there.

In the worst case, we must reconsider this dependency, as it
might become a window of opportunity to writing on the filesystem,
even if on a very specific folder.