Don't store path to the source tree in binaries
The tracker-needle & tracker-preferences tools had a developer
convenience that would hardcode the path to the source tree into the
binaries, so that while developing you could edit the .ui files without
having to rerun `make install` afterwards.

It's not good to enable this sort of thing in non-debug binaries, and
also it prevents the binaries from being bit-for-bit reproducible.

This problem was detected by the Debian reproducible builds CI system:

   https://tests.reproducible-builds.org/tracker