ontologies: Build dc/xsd/rdf/nrl/nao as GResource
Those ontology files are the tracker core, as the database files' structure
is modeled around those classes and properties. Include those as a GResource
so they're guaranteed to be there, all other ontology files are optional
and must be provided on TrackerDataManager initialization.

The NAO ontology has been split between the official bits (included in the
GResource) and the non-official bits (just used by the evolution miner, and
moved to 90-tracker)