libtracker-miner: Change API of ::process-file[-attributes]
There's a few changes here:
- The 2 vmethods are now given a GTask, its cancellable is to be
  used if the handling goes async.
- tracker_miner_fs_file_notify() has changed into a more generic
  tracker_miner_fs_notify_finish() method, that takes such GTask
  and completes it.
- The vmethods are no longer given a TrackerSparqlBuilder, instead
  they are expected to create the SPARQL through whatever mean is
  most fit. The sparql is given in the tracker_miner_fs_notify_finish()
  func. This opens the door to TrackerMinerFS implementations using
  TrackerResource.

The intent is 1) Pass something to these vmethods that the user
can't forge or mess with, as matching on GFile relies that it's the
same pointer that it was given in the vmethods. And 2) Make the finish()
function more generic to be fit to other methods going async.