libtracker-miner: Split ::remove-file into two signals
The only_children argument is a bit awkward as we emit ::remove-file
on a file that is not removed at all. The TrackerSparqlBuilder argument
has been also removed (the signals just have a gchar* return value
containing the SPARQL for the delete op) so it's up to the caller to
decide how to compose the SPARQL.

This allows removing some more knowledge about specific ontologies
from TrackerMinerFS, the ontology-dependent upper layers will know
better how to delete the corresponding entities.