libtracker-miner: Remove tracker_miner_fs_set_mtime_checking()
This is just used to set the TRACKER_DIRECTORY_FLAG_CHECK_MTIME flag on
the TrackerIndexingTree for all files. Given libtracker-miner has this
fine grained switch and all use of it happens in src/miners/fs, just move
the global toggle there and remove it from libtracker-miner API.

The only usage of this flag inside libtracker-miner happened inside
tracker_miner_fs_directory_add(), which was superseded by
TrackerIndexingTree too and is scheduled for removal.