tracker-miner-fs: Implement initial crawling fully here
It seems this setting was just ignored at on libtracker-miner since the
introduction of TrackerFileNotifier. And it does not make a lot of sense
either, crawling is done for 2 purposes: checking mtimes and setting up
monitors.

So we have individual toggles for each of those, but if neither happens
we could just not start the miner as well, which is most similar to the
tracker-miner-fs behavior when this setting was introduced.