Fix missing introspection data for libtracker-sparql (when using Meson)
Up til now only the Tracker.SparqlConnection and Tracker.SparqlBuilder
resources were introspectable. This is because we only used the
introspection output from `valac`, but other bits of libtracker-sparql
have since been added that are written in C.

There seems to be no way to generate a single .gir for a combined C and
Vala codebase, so instead I have written a simple `g-ir-merge` script
which can combine two different namespaces into a single .gir.

This is currently tested and working with the Meson build instructions.
It would be possible to implement this for the Autotools build
instructions as well.

https://bugzilla.gnome.org/show_bug.cgi?id=782091