libtracker-miner: Add filter_event vmethod
This is not a signal because external users of a TrackerMiner have no
business in modifying behavior at this level, this is reserved for
subclasses that presumably know what they are doing.

This vmethod is toggled for every event that gets received from the
TrackerFileNotifier, before the file gets to hit any processing queue.