Delete tracker-preferences
This is a mix of store and miner features which can't be assumed to go
together. It is better to leave configuration UI up to the environment,
in the GNOME case this is gnome-control-center search panel.

For settings that fall in the cracks, or for environments that don't
integrate Tracker search in their configuration UI, there's the rather
capable dconf-editor which will provide all toggles, and will explain
them better than any UI we may have.

In the future the CLI tools might gain some configuration subcommand,
but it will not be a graphical UI.