Delete tracker-needle
It was cool when Tracker search was barely integrated anywhere.
Nowadays, it's nothing but an aging showcase app (eg. was conceived
in the gtk2 days and barely ported to gtk3, and it shows everywhere).

If you want a modern and capable search UI, try Nautilus. If you
are outraged that tracker-needle is gone and ended up here, by all
means be my guest, reset to the parent of this commit and take it
out of tree. If you join #tracker I'll gladly give you advise on
how could it be brought up-to-date. But our hands are full, sorry,
we leave cool UIs to others.