meson: Rearrange dependencies until things compile
Targets far down the dependency chain were failing due to mysteriously
missing symbols. The use of -Wl,--as-needed seemed to be triggering the
issue.

This only occured with new versions of Meson. Previously circular
dependencies would be added recursively but this was removed with
<https://github.com/mesonbuild/meson/pull/1545>.

I'm not 100% sure how this fixes things, but I think that removing
or at least reshuffling circular dependencies of libtracker-sparql
means that libraries are passed to the linker in "correct" order.