libtracker-data: Refactor Resource table creation
Instead of creating it just in time for the rdfs:Resource table,
create it in advance so we have a common hook for other essential
tables.

The "Available" column that we add to the main class is kind of
useless, but is preserved so older Tracker versions don't stumble
on a database created with a more modern Tracker.

We are also adding a "Refcount" column to the Resource table, which
will be used in the following commits.