libtracker-data: Add resource element triggers
Those will control the Refcount column in the Resource table,
whenever a resource is added/removed from a resource or multivalued
property table, or whenever a single or multivalued property points
or stops pointing to a given resource, its refcount will be changed.

This adds automatic refcounting on elements, so we can tell those
with a refcount=0 are not referenced anywhere so thus can be dropped.