libtracker-data: Add Graph table
This new table will perform the accounting of those URNs used
as a graph on SPARQL updates. Those are kind of special in that
they are not referenced anywhere else, so we couldn't distinguish
between those and an entirely unused URN.

The creation of this new table has also been made to trigger
the recreation of all class tables, so the resource refcounting
changes in the previous commits become effective.