libtracker-miner: Call next_files_async() in batches
Instead of painstakingly requesting GFileInfos one by one,
request several of those in a single batch. This is a lot easier
to the main loop, because we don't need hitting it several times
per file.

This and the previous commits result in massive speedups to
TrackerMinerFS, since CPU time in userspace during TrackerMinerFS
activity was largely dominated by GSource and main loop overhead.
Somewhere around 1.80x-2x faster for initial indexing, and up to
3x faster for startup on an already indexed FS, from unscientifical
testing with a quarter million files contained in ~17K folders.